Clarify attribute definitions
Some minor tweaks to the syntax: attribute names are now given as
identifiers, not strings, and the entity that the attribute is being
added to appears in parentheses.

More substantially, you now use the `attr` statement to add attributes
to both graph nodes and edges.  The `node` and `edge` statements are
_only_ used to create create nodes and edges, and you must perform one
of those statements before using an `attr` statement to add an
attribute.

Lastly, this also removes the implicit merging of attribute contents.
We're not using this currently, and if we do need the ability to mutate
the value of an attribute, we should add all of the pieces to give the
DSL author full control over when and how to do so.