Remove Context object
In retrospect I think this was a premature optimization.  We have to
spend a decent amount of time hashing the identifier strings in order to
dedup them, and it's not at all obvious that that's less time than it
takes to just compare the (small) strings directly.  So now Identifiers
are just a wrapper around a reference-counted String.  This lets us get
rid of the Context object completely, greatly simplifying our Display
logic and several type signatures.

There is more that we could do here — for instance, allowing the content
of an Identifier to be a slice into the DSL file — but this is the
minimal change that gets rid of Context.