Add operator for testing whether a capture exists
Adds the prefix operator `?` which coerces capture
existence/non-existence into a boolean value. Thus, `? @foo` is `#true`
if the capture `@foo` exists, and is `#false` otherwise.

Because a capture that _doesn't_ exist results in an error if evaluated,
this behaviour cannot be implemented in terms of the existing
function call mechanism.

The syntax itself is also not entirely ideal. For ease of parsing, it's
better for it to be a prefix operator, but `?@foo` is not the prettiest
combination of symbols.

Alternatively, one might dispense with the `@` in this expression, and
write it simply as `?foo`, though that might cause confusion if there
is also a local variable called `foo` in the given stanza.