Fix execution of `print` statements
Because strings are (by default) printed with quotes, a statement such
as

```
print "x = ", 5
```

would result in the following being printed to `stdout`:

```
"x = "5
```

To make the output a bit nicer, we now do a case analysis on the value
being printed, and if it is a string already we simply emit that string
without escaping it.

No test case this time around, since executing the `print` statement
prints directly to `stdout`, and there didn't seem to be an easy way to
capture this output in a test.